Split logger into interface and implementation, reducing the number of visible headers
