+++ libosmscout-client-qt/src/osmscoutclientqt/ElevationChartWidget.cpp
@@ -18,10 +18,15 @@
  */
 
 #include <osmscoutclientqt/ElevationChartWidget.h>
-#include <osmscoutclientqt/OSMScoutQt.h>
-#include <osmscout/util/String.h>
 
 #include <QPainterPath>
+
+#include <limits>
+#include <sstream>
+
+#include <osmscout/util/String.h>
+
+#include <osmscoutclientqt/OSMScoutQt.h>
 
 namespace osmscout {
